backend/pyproject.toml
[tool.poetry]
name = "nupedee"
version = "0.1.0"
description = "Site novo do NUPEDEE"
authors = ["Vinicius Monego <monego@posteo.net>"]
license = "agpl3+"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
mysqlclient = "^2.1.1"
Pillow = "^9.4.0"
python-decouple = "^3.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
